requirements.txt
dj-database-url==0.5.0
Django==3.2.6
django-allauth==0.45.0
django-bootstrap-form==3.4
django-braces==1.14.0
django-ckeditor==6.0.0
django-crispy-forms==1.9.1
django-debug-toolbar==3.2.1
django-filter==2.2.0
django-google-maps==0.12.4
django-heroku==0.3.1
django-js-asset==1.2.2
django-model-utils==4.1.1
django-social-share==2.2.1
django-star-ratings==0.9.1
django-summernote==0.8.11.6
docutils==0.17.1
filelock==3.0.12
gunicorn==19.9.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
mysqlclient==1.4.6
numpy==1.18.2
oauthlib==3.1.1
twilio==6.63.0
urllib3==1.25.8
virtualenv==20.0.10
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
whitenoise==4.1.2
wrapt==1.12.1
